ai-service/requirements.txt
fastapi
uvicorn
pandas
chromadb
sentence-transformers
google-generativeai
python-multipart
pydantic
